remove prnt warning

+++ boot_control/Src/stack_config.c
@@ -93,7 +93,7 @@ void traverse_stack(void)
     const size_t n_words = copy_bytes / sizeof(uint32_t);
     if (size_bytes > STACK_USAGE_BYTES)
     {
-        printf("WARN: Traversing only the first %u bytes of the stack", copy_bytes);
+        printf("WARN: Traversing only the first %u bytes of the stack", (unsigned) copy_bytes);
     }
 
     memcpy(stack_copy, &_sstack, copy_bytes);
